Windows 7 support

+++ src/D3D9/texmgr.h
@@ -87,7 +87,7 @@ extern HMODULE d3dx9_dll;
     void                     reset (void);
     void                     purge (void); // WIP
 
-    int                      numTextures (void) {
+    size_t                   numTextures (void) {
       return textures.size ();
     }
     int                      numInjectedTextures (void);
