+++ src/parameter.cpp
@@ -131,8 +131,8 @@ sk::ParameterFloat::get_value_str (void)
   swprintf (val_str, L"%f", value);
 
   // Remove trailing 0's after the .
-  int len = wcslen (val_str);
-  for (int i = (len - 1); i > 1; i--) {
+  size_t len = wcslen (val_str);
+  for (size_t i = (len - 1); i > 1; i--) {
     if (val_str [i] == L'0' && val_str [i - 1] != L'.')
       len--;
     if (val_str [i] != L'0' && val_str [i] != L'\0')
